Changed some designs and put all the files in their folders respectively.
