--- src/App.css
@@ -1,185 +1,0 @@
-.App {
-  text-align: center;
-  background-color: #282c34;
-  height: 950px;
-  background-image: url(background1.jpg);
-  width: 100%;
-}
-
-.aiversebg{
-  height: 950px;
-  background-image: url(background1.jpg);
-  width: 100%;
-}
-
-/* .App-header {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-} */
-
-.App-link {
-  color: #61dafb;
-}
-
-.header-aiverse{
-  /* background: linear-gradient(90deg, #FF6706 -6.06%, #0048F5 56.97%);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent; */
-  background-clip: text;
-  font-style: normal;
-  font-weight: 800;
-  color: #FFFFFF;
-  width: 40%;
-  font-size: 4rem;
-}
-
-.top-header{
-  display: flex;
-  text-align: left;
-  margin-left: 13%;
-}
-
-.box{
-  height: 96px;
-  width: 29%;
-  background: #7b71a05c;
-  border-top-right-radius: 50px;
-  margin-left: 9%;
-  margin-top: -2%;
-}
-
-.p1{
-  font-family: 'Inter';
-  font-style: normal;
-  font-weight: 300;
-  font-size: 12px;
-  line-height: 160%;
-  color: white;
-  margin-top: 17%;
-  margin-left: -40%;
-}
-
-.robot{
-  display: flex;
-  width: 378px;
-  transform: scaleX(-1);
-  height: 50%;
-  margin-top: -15%;
-  margin-left: 10%;
-}
-
-h3{
-    margin-left: 15%;
-    width: 40%;
-    height: 90px;
-    font-family: 'Orbitron', sans-serif;
-    font-weight: 600;
-    font-size: 30px;
-    line-height: 40px;
-    display: flex;
-    align-items: center;
-    text-align: left;
-    color: #FFFFFF;
-}
-
-.space{
-    width: 25%;
-    height: 75%;
-    position: relative;
-    top: -105px;
-    left: 26%;
-}
-
-.techImg{
-  width: 16%;
-  height: 62%;
-  transform: rotate(40deg);
-  object-fit: unset;
-  size: landscape;
-  margin-left: 50%;
-  margin-top: -19%;
-}
-
-
-@media only screen and (max-width: 1400px){
-    .header-aiverse{
-      width: 45%;
-    }
-    .h2{
-      width: 50%;
-      font-size: 60px;
-    }
-    .techImg{
-      width: 20%;
-    }
-}
-
-@media only screen and (max-width: 1024px){
-  .header-aiverse{
-    width: 55%;
-    font-size: 3.5rem;
-  }
-  .p1{
-    margin-top: 23%;
-    margin-left: -55%;
-  }
-  .techImg{
-    width: 22%;
-    transform: rotate(35deg);
-  }
-  .box{
-    width: 40%;
-  }
-  .h2, .p2{
-    display: none;
-  }
-}
-
-@media only screen and (max-width: 800px){
-  .box{
-    width: 50%;
-  }
-  .robot{
-    width: 310px;
-  }
-  .techImg{
-    width: 28%;
-    transform: rotate(35deg);
-  }
-  .top-header{
-    width: 120%;
-    margin-left: 40%;
-  }
-}
-
-@media only screen and (max-width: 650px){
-  .App{
-    width: 130%;
-  }
-  .header-aiverse{
-    width: 70%;
-    text-align: center;
-    font-size: 2rem;
-  }
-  .p1 {
-    margin-top: 13%;
-    margin-left: -61%;
-}
-  .robot{
-    width: 280px;
-  }
-  .box{
-    width: 70%;
-    margin-left: -30%;
-  }
-  .techImg{
-    width: 44%;
-    height: 70%;
-    transform: rotate(37deg);
-  }
-}
