+++ src/assets/images/logo4.png
