+++ src/components/Engage/Engage.css
@@ -1,56 +1,55 @@
 .engagePage {
-    background-color: rgb(51, 10, 85);
+    background-color: black;
     width: 100%;
     height: 100vh;
-display: flex;
-align-items: center;
-justify-content: center;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
 
-
+.engageFilter {
+    width: 50%;
+    position: absolute;
+    height: 50%;
+    top: 250%;
+    background: linear-gradient(180deg, #2603FB 0%, #DC06FF 100%);
+    filter: blur(500px);
+    transform: matrix(0.93, -0.41, 0.36, 0.92, 0, 0);
 }
 
 .engageParent {
-
-    width: 1235px;
+    width: 100%;
     height: 517px;
-
 }
 
 .engageTitle {
-
-    /* position: absolute; */
+    text-align: left;
     width: 600px;
     height: 186.48px;
     margin-bottom: 88.52px;
-    font-family: 'Inter';
     font-style: normal;
     font-weight: 600;
-    font-size: 80px;
-
+    font-size: 50px;
     color: #FFFFFF;
-
-
-
+    margin-top: 0;
+    margin-left: 2%;
 }
 
 .engageSteps {
-    /* position: absolute; */
     width: 342px;
     height: 242px;
     margin-left: 52px;
     border-radius: 0px 0px 60px 0px;
-
     color: #FFFFFF;
     box-sizing: border-box;
-
     background: linear-gradient(90deg, rgba(255, 255, 255, 0.1) -1.52%, rgba(255, 255, 255, 0.024) 104.35%);
+    backdrop-filter: blur(120px);background: linear-gradient(90deg, rgba(255, 255, 255, 0.1) -1.52%, rgba(255, 255, 255, 0.024) 104.35%);
     backdrop-filter: blur(120px);
-
-
 }
 
 .engageStepBox {
     display: flex;
+    justify-content: center;
 }
 
 .engageStepNum {
@@ -64,13 +63,19 @@ justify-content: center;
 }
 
 .engageStepName {
-
-
     font-family: 'Inter';
     font-style: normal;
     font-weight: 300;
     font-size: 36px;
     margin: 35px;
     color: #FFFFFF;
-
 }
+
+.skull{
+    position: absolute;
+    top: 305%;
+    width: 15%;
+    right: 10%;
+    border-top-left-radius: 80px;
+    border-bottom-right-radius: 80px;
+  }
