+++ src/components/Header/Header.css
@@ -5,6 +5,7 @@
     flex-direction: column;
     width: 100%;
     height: 100vh;
+    background-color: black;
 }
 
 .headerEllipse {
@@ -59,13 +60,13 @@
 }
 
 .headerLogo {
-    width: 100px;
+    width: 90px;
     height: 50px;
 }
 
 .headerTitles {
     display: flex;
-    justify-content: center;
+    justify-content: flex-start;
     align-items: flex-start;
     flex-direction: column;
 }
@@ -74,6 +75,7 @@
     font-weight: bold;
     color: white;
     margin: 0;
+    text-align: start;
 }
 
 .nfts {
@@ -134,16 +136,19 @@
     font-size: 80px;
     margin-bottom: 0;
     font-weight: bold;
+    text-align: start;
 }
 
 .headerSubHeading {
     font-size: 60px;
     margin: 8px 0;
+    text-align: start;
 }
 
 .headerHeadingContent {
     margin: 0;
     line-height: 1.5;
+    text-align: start;
 }
 
 .poweredBy {
@@ -194,11 +199,6 @@
     }
     
     .headerEllipseTwo {
-        /* width: 300px;
-        height: 250px;
-        top: auto;
-        left: auto;
-        transform: none; */
         display: none;
     }
 
@@ -228,7 +228,7 @@
         width: 90%;
         justify-content: center;
         align-items: center;
-        background-image: url(landinghand.png);
+        background-image: url('../../assets/images/landinghand.png');
         background-repeat: no-repeat;
         background-position: center;
         background-size: 80% 125%;
