+++ src/components/Mint/Mint.css
@@ -44,6 +44,7 @@
     width: 40%;
     background: linear-gradient(90deg, rgba(0, 0, 0, 0.38) -1.46%, rgba(39, 31, 28, 0.0798) 100%);
     backdrop-filter: blur(14px);
+    cursor: pointer;
 }
 
 
@@ -124,6 +125,7 @@
     outline: none;
     padding: 15px 15px;
     width: 35%;
+    cursor: pointer;
 }
 
 .mintEmailButton {
